requirements.txt
beautifulsoup4==4.12
pypub3==2.0
requests==2.31
